add_document_folder

+++ app/src/main/java/oneBill/domain/entity/error/AmountMismatchException.java
@@ -4,7 +4,7 @@ package oneBill.domain.entity.error;
  * Created by 豪豪 on 2015/11/26.
  */
 public class AmountMismatchException extends MismatchException {
-    public AmountMismatchException(float amountpaid, float amountpayable){
+    public AmountMismatchException(double amountpaid, double amountpayable){
         super("应付、实付不匹配，总实付金额为" + amountpaid + "元，总应付金额为" + amountpayable + "元");
     }
 }
